requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
boto3==1.16.40
botocore==1.19.40
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
click==7.1.2
discord-interactions==0.1.0
discord.py==1.5.1
Flask==1.1.2
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
multidict==4.7.6
pycparser==2.20
PyNaCl==1.4.0
python-dateutil==2.8.1
python-dotenv==0.15.0
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.5.1
